test: Move remaining js-disabling tests to ExpandPlaceholders.test.ts
