+++ tests/Scripting/ExpandPlaceholders.test.ts
@@ -1,6 +1,43 @@
 import { expandPlaceholders } from '../../src/Scripting/ExpandPlaceholders';
 import { makeQueryContext } from '../../src/Scripting/QueryContext';
 import { TasksFile } from '../../src/Scripting/TasksFile';
+import { Query } from '../../src/Query/Query';
+import { EnableJsInTasksQueries } from '../../src/Config/EnableJsInTasksQueries';
+
+describe('Placeholders - disabling execution', () => {
+    beforeEach(() => {
+        EnableJsInTasksQueries.getInstance().set(false);
+    });
+
+    afterEach(() => {
+        EnableJsInTasksQueries.getInstance().set(true);
+    });
+
+    const tasksFile = new TasksFile('anywhere.md');
+
+    function checkQueryErrorMessage(query: Query, instruction: string): void {
+        expect(query.error).toContain(EnableJsInTasksQueries.getHelpMessage());
+        expect(query.error).toContain(instruction);
+    }
+
+    it('"{{query.file.path}}" should work when JS execution disabled', () => {
+        const instruction = 'path includes {{query.file.path}}';
+        const query = new Query(instruction, tasksFile);
+        expect(query.error).toBeUndefined();
+    });
+
+    it('"{{query.file.path.toUpperCase()}}" should have meaningful parse-time error', () => {
+        const instruction = 'path includes {{query.file.path.toUpperCase()}}';
+        const query = new Query(instruction, tasksFile);
+        checkQueryErrorMessage(query, instruction);
+    });
+
+    it.failing('"{{4 + 6}}" should have meaningful parse-time error', () => {
+        const instruction = 'path includes {{4 + 6}}';
+        const query = new Query(instruction, tasksFile);
+        checkQueryErrorMessage(query, instruction);
+    });
+});
 
 describe('ExpandTemplate', () => {
     const tasksFile = new TasksFile('a/b/path with space.md');
